requirements.txt
absl-py==0.6.1
aes==1.0.0
aiohttp==3.6.2
arff==0.9
astor==0.7.1
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
blessings==1.7
boto3==1.13.1
botocore==1.16.1
camelot-py==0.7.3
catboost==0.11.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
ci-info==0.2.0
click==7.1.1
configobj==5.0.6
configparser==5.0.0
crypto==1.4.1
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
elasticsearch==7.6.0
Elixir==0.7.1
enum34==1.1.6
et-xmlfile==1.0.1
etelemetry==0.2.1
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
fitz==0.0.1.dev2
Flask==1.1.2
ftransc==7.0.1
future==0.18.2
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
html2text==2020.1.16
httplib2==0.17.3
idna==2.6
idna-ssl==1.1.0
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.15.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.5.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.2.1
multidict==4.7.5
mutagen==1.44.0
Naked==0.1.31
networkx==2.4
neurdflib==5.0.1
newspaper3k==0.2.8
nibabel==3.1.0
nipype==1.4.2
nltk==3.5
numpy==1.14.3
objectpath==0.6.1
octopart==0.0.24
opencv-python==4.1.2.30
openpyxl==3.0.3
packaging==20.3
pafy==0.5.5
pandas==0.23.4
pbr==5.4.5
pdfminer==20191125
pdfminer.six==20200402
Pillow==7.1.2
plumbum==1.6.9
protobuf==3.6.1
prov==1.5.3
pycparser==2.20
pycryptodome==3.9.1
pycryptodomex==3.9.1
pydot==1.4.1
pydotplus==2.0.2
PyMuPDF==1.16.18
pyOpenSSL==19.0.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyTelegramBotAPI==3.6.7
pytesseract==0.3.4
python-dateutil==2.7.3
python-dotenv==0.13.0
python-telegram-bot==12.6.1
pytz==2018.4
pyxnat==1.3
PyYAML==3.13
rdflib==5.0.0
readability-lxml==0.7.1
regex==2020.5.14
requests==2.23.0
requests-file==1.5.1
resp==0.1.2
retrying==1.3.3
s3transfer==0.3.3
schematics==2.1.0
scikit-learn==0.20.0
scipy==1.1.0
shellescape==3.4.1
simplejson==3.17.0
six==1.14.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==2.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
telebot==0.0.3
telepot==12.7
Tempita==0.5.2
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
tinysegmenter==0.3
tldextract==2.2.2
tornado==6.0.4
tqdm==4.46.0
traits==6.0.0
typing-extensions==3.7.4.2
urllib3==1.22
validators==0.14.3
Werkzeug==1.0.1
xlrd==0.7.1
xlwt==0.7.2
yarl==1.4.2
youtube-dl==2020.3.24
